# Frontend/RepoGeist/openapi.json
{
    "openapi":  "3.1.1",
    "info":  {
                 "title":  "RepoGeist | v1",
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:5221/"
                    }
                ],
    "paths":  {
...
                  "/api/repos/{owner}/{name}/clone-status":  {
                                                                 "get":  {
                                                                             "tags":  [
                                                                                          "Clone"
                                                                                      ],
                                                                             "parameters":  [
                                                                                                {
                                                                                                    "name":  "owner",
                                                                                                    "in":  "path",
                                                                                                    "required":  true,
                                                                                                    "schema":  {
                                                                                                                   "type":  "string"
                                                                                                               }
                                                                                                },
                                                                                                {
                                                                                                    "name":  "name",
                                                                                                    "in":  "path",
                                                                                                    "required":  true,
                                                                                                    "schema":  {
                                                                                                                   "type":  "string"
                                                                                                               }
                                                                                                }
                                                                                            ],
                                                                             "responses":  {
                                                                                               "200":  {
                                                                                                           "description":  "OK",
                                                                                                           "content":  {
                                                                                                                           "text/plain":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/CloneStatusDto"
                                                                                                                                                         }
                                                                                                                                          },
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/CloneStatusDto"
                                                                                                                                                               }
                                                                                                                                                },
                                                                                                                           "text/json":  {
                                                                                                                                             "schema":  {
                                                                                                                                                            "$ref":  "#/components/schemas/CloneStatusDto"
                                                                                                                                                        }
                                                                                                                                         }
                                                                                                                       }
                                                                                                       },
                                                                                               "404":  {
                                                                                                           "description":  "Not Found",
                                                                                                           "content":  {
                                                                                                                           "text/plain":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                         }
                                                                                                                                          },
                                                                                                                           "application/json":  {
                                                                                                                                                    "schema":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                               }
                                                                                                                                                },
                                                                                                                           "text/json":  {
                                                                                                                                             "schema":  {
                                                                                                                                                            "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                        }
                                                                                                                                         }
                                                                                                                       }
                                                                                                       }
                                                                                           }
                                                                         }
                                                             },
                  "/api/repos/{owner}/{name}/file-tree":  {
                                                              "get":  {
                                                                          "tags":  [
                                                                                       "Clone"
                                                                                   ],
                                                                          "parameters":  [
                                                                                             {
                                                                                                 "name":  "owner",
                                                                                                 "in":  "path",
                                                                                                 "required":  true,
                                                                                                 "schema":  {
                                                                                                                "type":  "string"
                                                                                                            }
                                                                                             },
                                                                                             {
                                                                                                 "name":  "name",
                                                                                                 "in":  "path",
                                                                                                 "required":  true,
                                                                                                 "schema":  {
                                                                                                                "type":  "string"
                                                                                                            }
                                                                                             }
                                                                                         ],
                                                                          "responses":  {
                                                                                            "200":  {
                                                                                                        "description":  "OK",
                                                                                                        "content":  {
                                                                                                                        "text/plain":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "$ref":  "#/components/schemas/FileTreeDto"
                                                                                                                                                      }
                                                                                                                                       },
                                                                                                                        "application/json":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/FileTreeDto"
                                                                                                                                                            }
                                                                                                                                             },
                                                                                                                        "text/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/FileTreeDto"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                                    }
                                                                                                    },
                                                                                            "404":  {
                                                                                                        "description":  "Not Found",
                                                                                                        "content":  {
                                                                                                                        "text/plain":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                      }
                                                                                                                                       },
                                                                                                                        "application/json":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                            }
                                                                                                                                             },
                                                                                                                        "text/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/ProblemDetails"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                                    }
                                                                                                    }
                                                                                        }
                                                                      }
                                                          },
...
                  "/api/repos/{owner}/{name}":  {
                                                    "get":  {
                                                                "tags":  [
                                                                             "Repos"
                                                                         ],
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "owner",
                                                                                       "in":  "path",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                   },
                                                                                   {
                                                                                       "name":  "name",
                                                                                       "in":  "path",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "OK",
                                                                                              "content":  {
                                                                                                              "text/plain":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/RepoDetailDto"
                                                                                                                                            }
                                                                                                                             },
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/RepoDetailDto"
                                                                                                                                                  }
                                                                                                                                   },
                                                                                                              "text/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/RepoDetailDto"
                                                                                                                                           }
                                                                                                                            }
                                                                                                          }
                                                                                          }
                                                                              }
                                                            }
                                                },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CloneStatusDto":  {
                                                              "type":  "object",
                                                              "properties":  {
                                                                                 "repoId":  {
                                                                                                "type":  "string",
                                                                                                "format":  "uuid"
                                                                                            },
                                                                                 "jobId":  {
                                                                                               "type":  "string",
                                                                                               "format":  "uuid"
                                                                                           },
                                                                                 "status":  {
                                                                                                "type":  "string"
                                                                                            },
                                                                                 "progressPercent":  {
                                                                                                         "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                                         "type":  [
                                                                                                                      "null",
                                                                                                                      "integer",
                                                                                                                      "string"
                                                                                                                  ],
                                                                                                         "format":  "int32"
                                                                                                     },
                                                                                 "errorMessage":  {
                                                                                                      "type":  [
                                                                                                                   "null",
                                                                                                                   "string"
                                                                                                               ]
                                                                                                  }
                                                                             }
                                                          },
...
                                   }
                   },
...
}